Rename app BarteguidenApp => Barteguiden
Must be done, since the project otherwise will be displayed as BarteguidenApp to users. There was a simpler way for android, but for iOS the project had to be renamed, and the cleaner solution then was to rename the project in package.json and run react-native upgrade.
