+++ android/app/build.gradle
@@ -92,7 +92,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.barteguidenapp"
+        applicationId "com.barteguiden"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
