+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'BarteguidenApp'
+rootProject.name = 'Barteguiden'
 
 include ':app'
