+++ index.ios.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
 import BarteguidenApp from './src/containers/App';
 
-AppRegistry.registerComponent('BarteguidenApp', () => BarteguidenApp);
+AppRegistry.registerComponent('Barteguiden', () => BarteguidenApp);
