+++ ios/Barteguiden.xcodeproj/xcshareddata/xcschemes/Barteguiden.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "BarteguidenApp.app"
-               BlueprintName = "BarteguidenApp"
-               ReferencedContainer = "container:BarteguidenApp.xcodeproj">
+               BuildableName = "Barteguiden.app"
+               BlueprintName = "Barteguiden"
+               ReferencedContainer = "container:Barteguiden.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BarteguidenAppTests.xctest"
-               BlueprintName = "BarteguidenAppTests"
-               ReferencedContainer = "container:BarteguidenApp.xcodeproj">
+               BuildableName = "BarteguidenTests.xctest"
+               BlueprintName = "BarteguidenTests"
+               ReferencedContainer = "container:Barteguiden.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -47,9 +47,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BarteguidenAppTests.xctest"
-               BlueprintName = "BarteguidenAppTests"
-               ReferencedContainer = "container:BarteguidenApp.xcodeproj">
+               BuildableName = "BarteguidenTests.xctest"
+               BlueprintName = "BarteguidenTests"
+               ReferencedContainer = "container:Barteguiden.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -57,9 +57,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BarteguidenApp.app"
-            BlueprintName = "BarteguidenApp"
-            ReferencedContainer = "container:BarteguidenApp.xcodeproj">
+            BuildableName = "Barteguiden.app"
+            BlueprintName = "Barteguiden"
+            ReferencedContainer = "container:Barteguiden.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -80,9 +80,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BarteguidenApp.app"
-            BlueprintName = "BarteguidenApp"
-            ReferencedContainer = "container:BarteguidenApp.xcodeproj">
+            BuildableName = "Barteguiden.app"
+            BlueprintName = "Barteguiden"
+            ReferencedContainer = "container:Barteguiden.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -99,9 +99,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BarteguidenApp.app"
-            BlueprintName = "BarteguidenApp"
-            ReferencedContainer = "container:BarteguidenApp.xcodeproj">
+            BuildableName = "Barteguiden.app"
+            BlueprintName = "Barteguiden"
+            ReferencedContainer = "container:Barteguiden.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
