+++ ios/Barteguiden/Images.xcassets/AppIcon.appiconset/Contents.json
