+++ ios/Barteguiden/Images.xcassets/AppIcon.appiconset/Icon-Small@2x.png
