+++ ios/Barteguiden/Images.xcassets/AppIcon.appiconset/Icon@2x-1.png
