+++ ios/Barteguiden/Images.xcassets/Contents.json
