+++ ios/Barteguiden/Images.xcassets/all.imageset/Contents.json
