+++ ios/Barteguiden/Images.xcassets/all.imageset/all@2x.png
