+++ ios/Barteguiden/Images.xcassets/favorites.imageset/Contents.json
