+++ ios/Barteguiden/Images.xcassets/favorites.imageset/favorites.png
