+++ ios/Barteguiden/Images.xcassets/favorites.imageset/favorites@2x.png
