+++ ios/Barteguiden/Images.xcassets/settings.imageset/Contents.json
