+++ ios/Barteguiden/Images.xcassets/settings.imageset/settings.png
