+++ ios/Barteguiden/Images.xcassets/settings.imageset/settings@2x.png
