+++ ios/BarteguidenTests/BarteguidenTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface BarteguidenAppTests : XCTestCase
+@interface BarteguidenTests : XCTestCase
 
 @end
 
-@implementation BarteguidenAppTests
+@implementation BarteguidenTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
