+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "BarteguidenApp",
+  "name": "Barteguiden",
   "version": "0.0.1",
   "private": true,
   "scripts": {
